Update `FragmentParser` to handle tables and unsupported elements, enhance placeholder and paragraph parsing, and refactor image handling in runs. Add `GenericNode` and `ImageNode` for extensible DOM representation. Improve `DocxEmitter` for rendering tables, inline images, and nested nodes.

+++ composeApp/src/commonMain/kotlin/dev/heizer/core/document/renderer/docx/model/DocModels.kt
@@ -26,8 +26,8 @@ data class RunNode(
 data class ImageNode(
     val data: ByteArray,
     val extension: String,
-    val width: Double,
-    val height: Double
+    val width: Long,
+    val height: Long
 ) : DocNode() {
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
